Moved type to types; fixed typos
