+++ src/app/types/location.type.ts
@@ -1,0 +1,11 @@
+type SetLocationValue = {
+  latitude: number,
+  longitude: number,
+  accuracy?: number,
+  altitude?: number,
+  altitudeAccuracy?: number,
+  heading?: number,
+  speed?: number
+}
+
+export { SetLocationValue }
